% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collect_variables}
\alias{collect_variables}
\title{Utility function to collect all the variable names from an open netCDF file.}
\usage{
collect_variables(nc_obj, nodes = FALSE)
}
\arguments{
\item{nc_obj}{An open fvcom netCDF file to extract the data from.}

\item{nodes}{For a fvcom format netCDF file, whether to return if the variable
makes use of the nodes or elements.}
}
\value{
A tibble of variable names, longnames and (if specified) nodes or
  elements.
}
\description{
Utility function to collect all the variable names from an open netCDF file.
}
\examples{
\dontrun{

 the_nc_obj <- ncdf4::nc_open("path/to/netCDF/file.nc")

 collect_variables(the_nc_obj, nodes = TRUE)
}
}
