% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{extract_site_grids_nc}
\alias{extract_site_grids_nc}
\title{Extract all variable values from netCDF file for all provided geographical locations.}
\usage{
extract_site_grids_nc(
  nc_file,
  sites,
  nc_crs,
  sites_crs = NULL,
  use_bbox = FALSE,
  depth_var = "depth",
  depths_to_extract = NULL,
  id = "site",
  nc_var_name = NULL,
  dates_to_extract = NULL,
  time_present = TRUE,
  time_var_name = "time",
  time_value = NULL,
  grid_name_x = "",
  grid_name_y = "",
  site_message = TRUE
)
}
\arguments{
\item{nc_file}{Filepath for the netCDF file to extract the data from.}

\item{sites}{Simple features (or spatial) data frame of geographical
areas to extract the variable values for. The data frame should contain a
single row per buffer and have a unique identifier.}

\item{nc_crs}{The numerical value for the crs of the netCDF file.}

\item{sites_crs}{The numerical value for the crs of sites, if different
from the crs of the netCDF file.}

\item{use_bbox}{A boolean value to choose whether the boundary box of the
sites will be used to speed up the extraction of data from the
netCDF file. When a netCDF file cannot be reduced to a small area, or is
particularly detailed, this can drastically speed up data extraction.}

\item{depth_var}{The name of the depth variable, if necessary, used in the
netCDF file. Defaults to "depth".}

\item{depths_to_extract}{The specific depth values to extract, if applicable.
By default, if any depth values are present, all of them will be extracted.}

\item{id}{The name of the identifier provided in the sites special
features data frame.}

\item{nc_var_name}{The name of the variable in the netCDF file that is the focus
of the extraction. If not specified, an attempt will be made to establish
this from the netCDF file itself.}

\item{dates_to_extract}{Vector of dates to extract the variable for. If not
provided, all available values will be returned.}

\item{time_present}{Boolean value that indicates whether a time, or similar
value is present in the netCDF file.}

\item{time_var_name}{Name of time/date variable to extract from the netCDF file.
Defaults to "time" as that is by far the most common variable name used.}

\item{time_value}{If there is no date/time value present in the netCDF file, a user
can provide the date(s) or time(s) for this netCDF file.}

\item{grid_name_x}{Name of the geographical x coordinate to be extracted.
If not specified, an attempt will be made to establish this from the netCDF
file itself.}

\item{grid_name_y}{Name of the geographical y coordinate to be extracted.
If not specified, an attempt will be made to establish this from the netCDF
file itself.}

\item{site_message}{Boolean variable to indicate whether to print the messages
about which site is being processed.}
}
\value{
A tibble of each location id, with the corresponding variable values
  depths and dates available.
}
\description{
Extract all variable values from netCDF file for all provided geographical locations.
}
\examples{
\dontrun{

 the_nc_file <- "path/to/netCDF/file.nc"

 # presuming we already have a set of spatial buffers to collect data
 # from
 extract_site_grids_nc(the_nc_file, sites = buffers, nc_crs = 4326)
}
}
