% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{filter_dates}
\alias{filter_dates}
\title{Internal function to assist the user in extracting only particular dates
from the results of an nc file.}
\usage{
filter_dates(all_nc_data, dates_to_extract)
}
\arguments{
\item{all_nc_data}{Data frame containing all the results extracted from an
nc file.}

\item{dates_to_extract}{Vector of dates to be extracted.}
}
\value{
Data frame containing the results filtered by date.
}
\description{
Internal function to assist the user in extracting only particular dates
from the results of an nc file.
}
