% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satpoint-package.R
\docType{package}
\name{satpoint-package}
\alias{satpoint}
\alias{satpoint-package}
\title{satpoint: Automates Extraction of Information From NETCDF Files}
\description{
The satpoint package is designed to make extracting information from NetCDF files at specific geographic locations as easy as possible. You provide some geographic data and an NetCDF file, satpoint will extract the variable data automatically.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://decide-project-eu.github.io/satpoint/}
}

}
\author{
\strong{Maintainer}: Andrew Duncan \email{andrew.duncan.ic@uhi.ac.uk} (\href{https://orcid.org/0000-0002-7314-9629}{ORCID})

Authors:
\itemize{
  \item Annette Boerlage \email{annette.boerlage@sruc.ac.uk} [contributor]
}

Other contributors:
\itemize{
  \item This work has received funding from the European Union’s Horizon 2020 research and innovation programme under grant agreement No. 101000494 (DECIDE), the Sustainable Aquaculture Innovation Centre (SAIC) under grant reference No. SL_2021_06, and the the Biotechnology and Biological Sciences Research Council (BBSRC) grant reference No. BB/Z515231/1. [funder]
}

}
\keyword{internal}
