% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffers.R
\name{site_buffers}
\alias{site_buffers}
\title{Creates circular buffers around site locations to allow for different
spatial averaging to occur when the values are extracted from an nc file.}
\usage{
site_buffers(
  site_df,
  x_coord,
  y_coord,
  crs,
  buffers,
  unite = FALSE,
  id = "site"
)
}
\arguments{
\item{site_df}{Data frame of locations to create buffers from. There is a
presumption that each location has a (x, y) coordinates in separate columns
and an identifier.}

\item{x_coord}{The name of the data frame column that details the x
coordinate of the geographical location.}

\item{y_coord}{The name of the data frame column that details the y
coordinate of the geographical location.}

\item{crs}{The numerical value of the to be used in the creation of the
geographical buffer zones. This should match the form of the coordinates
provided in the data frame.}

\item{buffers}{The radius of the buffers to be created in km.}

\item{unite}{Whether to unit the identifier of the location and the buffer.}

\item{id}{The column name that details the location identifier.
Defaults to "site".}
}
\value{
Special features data frame with the original data frame's column,
  plus a geometry column identifying the circular buffers around the
  individual locations.
}
\description{
Creates circular buffers around site locations to allow for different
spatial averaging to occur when the values are extracted from an nc file.
}
