% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{extract_dates}
\alias{extract_dates}
\title{Extract all date or datetimes from an nc file.}
\usage{
extract_dates(nc_obj, time_var_name = "time")
}
\arguments{
\item{nc_obj}{Open nc file to retrieve time data from}

\item{time_var_name}{Name of time/date variable to extract from the nc file.
Defaults to "time" as that is by far the most common variable name used.}
}
\value{
Vector of date or datetimes extracted from the nc file.
}
\description{
Extract all date or datetimes from an nc file.
}
