% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos.R
\name{extract_geos}
\alias{extract_geos}
\title{Extract all geographical coordinates from an nc file and convert to simple
  features object.}
\usage{
extract_geos(nc_obj, orig_crs, crs = NULL, name_x = "", name_y = "")
}
\arguments{
\item{nc_obj}{Open nc file to retrieve time data from.}

\item{orig_crs}{Numerical value of the crs used in the nc file.}

\item{crs}{Numerical value of the crs to change results into, if different
from the original crs.}

\item{name_x}{Name of the geographical x coordinate to be extracted.
If not specified, an attempt will be made to establish this from the nc
file itself.}

\item{name_y}{Name of the geographical y coordinate to be extracted.
If not specified, an attempt will be made to establish this from the nc
file itself.}
}
\value{
Special features data frame containing the geographic coordinates of
  the provided nc file.
}
\description{
Extract all geographical coordinates from an nc file and convert to simple
  features object.
}
