% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{extract_site_grids_nc}
\alias{extract_site_grids_nc}
\title{Extract all variable values from nc file for all provided geographical locations.}
\usage{
extract_site_grids_nc(
  nc_file,
  site_locations,
  nc_crs,
  sites_crs = NULL,
  depth_var = "depth",
  id = "site",
  nc_var_name = NULL,
  dates_to_extract = NULL,
  time_var_name = "time",
  grid_name_x = "",
  grid_name_y = ""
)
}
\arguments{
\item{nc_file}{File to extract the data from.}

\item{site_locations}{Simple features (or spatial data frame) of geographical
areas to extract the variable values for.}

\item{nc_crs}{The numerical value for the crs of the nc file.}

\item{sites_crs}{The numerical value for the crs of the sites, if different
from the crs of the nc file.}

\item{depth_var}{The name of the depth variable, if necessary, used in the
nc file. Defaults to "depth".}

\item{id}{The name of the identier provided in the sites_location special
features data frame.}

\item{nc_var_name}{The name of the variable in the nc file that is the focus
of the extraction. If not specified, an attempt will be made to establish
this from the nc file itself.}

\item{dates_to_extract}{Vector of dates to extract the variable for. If not
provided, all available values will be returned.}

\item{time_var_name}{Name of time/date variable to extract from the nc file.
Defaults to "time" as that is by far the most common variable name used.}

\item{grid_name_x}{Name of the geographical x coordinate to be extracted.
If not specified, an attempt will be made to establish this from the nc
file itself.}

\item{grid_name_y}{Name of the geographical y coordinate to be extracted.
If not specified, an attempt will be made to establish this from the nc
file itself.}
}
\value{
A tibble of each location id, with the corresponding variable values
  depths and dates available.
}
\description{
Extract all variable values from nc file for all provided geographical locations.
}
