% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{get_nc_data}
\alias{get_nc_data}
\title{Internal function to extract the variable values from an open nc file.}
\usage{
get_nc_data(
  ind.x,
  ind.y,
  nc_obj,
  depth_vals,
  nc_var,
  nc_times,
  site,
  swap_ind = FALSE
)
}
\arguments{
\item{ind.x}{Index of the geographic x coordinate to start the variable
extraction from.}

\item{ind.y}{Index of the geographic y coordinate to start the variable
extraction from.}

\item{nc_obj}{Open nc file to extract the data from.}

\item{depth_vals}{The depth values (if present) to match the extracted data
to.}

\item{nc_var}{Name of the variable contained in the nc file.}

\item{nc_times}{Vector of times contained with the nc file to match with the
extracted data.}

\item{site}{Name of the location that these values will be attached to.}

\item{swap_ind}{Boolean identifier of whether to change the order of the x, y
indices.}
}
\value{
Tibble of all variable values, times and depths extracted from the
  provided nc file.
}
\description{
Internal function to extract the variable values from an open nc file.
}
