% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satpoint-package.R
\docType{package}
\name{satpoint-package}
\alias{satpoint}
\alias{satpoint-package}
\title{satpoint: Automates Extraction Of Information From NETCDF Files}
\description{
The satpoint package is designed to make extracting information from NetCDF files at specific geographic locations as easy as possible. You provide some geographic data and an NetCDF file, satpoint will extract the variable data automatically.
}
\author{
\strong{Maintainer}: Andrew Duncan \email{Andrew.Duncan@sruc.ac.uk}

}
\keyword{internal}
